PLANNER-1688 Extract build methods from Config
This commit covers SolverConfig, PhaseConfig and ScoreDirectorFactoryConfig.
