+++ optaplanner-core/src/main/java/org/optaplanner/core/config/partitionedsearch/PartitionedSearchPhaseConfig.java
@@ -16,10 +16,8 @@
 
 package org.optaplanner.core.config.partitionedsearch;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.ThreadFactory;
 
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlElements;
@@ -33,26 +31,15 @@ import org.optaplanner.core.config.localsearch.LocalSearchPhaseConfig;
 import org.optaplanner.core.config.phase.NoChangePhaseConfig;
 import org.optaplanner.core.config.phase.PhaseConfig;
 import org.optaplanner.core.config.phase.custom.CustomPhaseConfig;
-import org.optaplanner.core.config.solver.EnvironmentMode;
 import org.optaplanner.core.config.util.ConfigUtils;
-import org.optaplanner.core.impl.heuristic.HeuristicConfigPolicy;
 import org.optaplanner.core.impl.io.jaxb.JaxbCustomPropertiesAdapter;
-import org.optaplanner.core.impl.partitionedsearch.DefaultPartitionedSearchPhase;
-import org.optaplanner.core.impl.partitionedsearch.PartitionedSearchPhase;
 import org.optaplanner.core.impl.partitionedsearch.partitioner.SolutionPartitioner;
-import org.optaplanner.core.impl.solver.recaller.BestSolutionRecaller;
-import org.optaplanner.core.impl.solver.termination.Termination;
-import org.optaplanner.core.impl.solver.thread.ChildThreadType;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class PartitionedSearchPhaseConfig extends PhaseConfig<PartitionedSearchPhaseConfig> {
 
     public static final String XML_ELEMENT_NAME = "partitionedSearch";
     public static final String ACTIVE_THREAD_COUNT_AUTO = "AUTO";
     public static final String ACTIVE_THREAD_COUNT_UNLIMITED = "UNLIMITED";
-
-    private static final Logger logger = LoggerFactory.getLogger(PartitionedSearchPhaseConfig.class);
 
     // Warning: all fields are null (and not defaulted) because they can be inherited
     // and also because the input config file should match the output config file
@@ -131,82 +118,6 @@ public class PartitionedSearchPhaseConfig extends PhaseConfig<PartitionedSearchP
 
     public void setPhaseConfigList(List<PhaseConfig> phaseConfigList) {
         this.phaseConfigList = phaseConfigList;
-    }
-
-    // ************************************************************************
-    // Builder methods
-    // ************************************************************************
-
-    @Override
-    public PartitionedSearchPhase buildPhase(int phaseIndex, HeuristicConfigPolicy solverConfigPolicy,
-            BestSolutionRecaller bestSolutionRecaller, Termination solverTermination) {
-        HeuristicConfigPolicy phaseConfigPolicy = solverConfigPolicy.createPhaseConfigPolicy();
-        ThreadFactory threadFactory = solverConfigPolicy.buildThreadFactory(ChildThreadType.PART_THREAD);
-        DefaultPartitionedSearchPhase phase = new DefaultPartitionedSearchPhase(
-                phaseIndex, solverConfigPolicy.getLogIndentation(), bestSolutionRecaller,
-                buildPhaseTermination(phaseConfigPolicy, solverTermination),
-                buildSolutionPartitioner(), threadFactory, resolvedActiveThreadCount());
-        List<PhaseConfig> phaseConfigList_ = phaseConfigList;
-        if (ConfigUtils.isEmptyCollection(phaseConfigList_)) {
-            phaseConfigList_ = Arrays.asList(
-                    new ConstructionHeuristicPhaseConfig(),
-                    new LocalSearchPhaseConfig());
-        }
-        phase.setPhaseConfigList(phaseConfigList_);
-        phase.setConfigPolicy(phaseConfigPolicy.createChildThreadConfigPolicy(ChildThreadType.PART_THREAD));
-        EnvironmentMode environmentMode = phaseConfigPolicy.getEnvironmentMode();
-        if (environmentMode.isNonIntrusiveFullAsserted()) {
-            phase.setAssertStepScoreFromScratch(true);
-        }
-        if (environmentMode.isIntrusiveFastAsserted()) {
-            phase.setAssertExpectedStepScore(true);
-            phase.setAssertShadowVariablesAreNotStaleAfterStep(true);
-        }
-        return phase;
-    }
-
-    private SolutionPartitioner buildSolutionPartitioner() {
-        if (solutionPartitionerClass != null) {
-            SolutionPartitioner<?> solutionPartitioner = ConfigUtils.newInstance(this,
-                    "solutionPartitionerClass", solutionPartitionerClass);
-            ConfigUtils.applyCustomProperties(solutionPartitioner, "solutionPartitionerClass",
-                    solutionPartitionerCustomProperties, "solutionPartitionerCustomProperties");
-            return solutionPartitioner;
-        } else {
-            if (solutionPartitionerCustomProperties != null) {
-                throw new IllegalStateException("If there is no solutionPartitionerClass (" + solutionPartitionerClass
-                        + "), then there can be no solutionPartitionerCustomProperties ("
-                        + solutionPartitionerCustomProperties + ") either.");
-            }
-            // TODO Implement generic partitioner
-            throw new UnsupportedOperationException();
-        }
-    }
-
-    private Integer resolvedActiveThreadCount() {
-        int availableProcessorCount = Runtime.getRuntime().availableProcessors();
-        Integer resolvedActiveThreadCount;
-        if (runnablePartThreadLimit == null || runnablePartThreadLimit.equals(ACTIVE_THREAD_COUNT_AUTO)) {
-            // Leave one for the Operating System and 1 for the solver thread, take the rest
-            resolvedActiveThreadCount = Math.max(1, availableProcessorCount - 2);
-        } else if (runnablePartThreadLimit.equals(ACTIVE_THREAD_COUNT_UNLIMITED)) {
-            resolvedActiveThreadCount = null;
-        } else {
-            resolvedActiveThreadCount = ConfigUtils.resolveThreadPoolSizeScript(
-                    "runnablePartThreadLimit", runnablePartThreadLimit, ACTIVE_THREAD_COUNT_AUTO,
-                    ACTIVE_THREAD_COUNT_UNLIMITED);
-            if (resolvedActiveThreadCount < 1) {
-                throw new IllegalArgumentException("The runnablePartThreadLimit (" + runnablePartThreadLimit
-                        + ") resulted in a resolvedActiveThreadCount (" + resolvedActiveThreadCount
-                        + ") that is lower than 1.");
-            }
-            if (resolvedActiveThreadCount > availableProcessorCount) {
-                logger.debug("The resolvedActiveThreadCount ({}) is higher than "
-                        + "the availableProcessorCount ({}), so the JVM will "
-                        + "round-robin the CPU instead.", resolvedActiveThreadCount, availableProcessorCount);
-            }
-        }
-        return resolvedActiveThreadCount;
     }
 
     @Override
