+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/partitionedsearch/DefaultPartitionedSearchPhaseFactory.java
@@ -1,0 +1,128 @@
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.optaplanner.core.impl.partitionedsearch;
+
+import static org.optaplanner.core.config.partitionedsearch.PartitionedSearchPhaseConfig.ACTIVE_THREAD_COUNT_AUTO;
+import static org.optaplanner.core.config.partitionedsearch.PartitionedSearchPhaseConfig.ACTIVE_THREAD_COUNT_UNLIMITED;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.concurrent.ThreadFactory;
+
+import org.optaplanner.core.config.constructionheuristic.ConstructionHeuristicPhaseConfig;
+import org.optaplanner.core.config.localsearch.LocalSearchPhaseConfig;
+import org.optaplanner.core.config.partitionedsearch.PartitionedSearchPhaseConfig;
+import org.optaplanner.core.config.phase.PhaseConfig;
+import org.optaplanner.core.config.solver.EnvironmentMode;
+import org.optaplanner.core.config.util.ConfigUtils;
+import org.optaplanner.core.impl.heuristic.HeuristicConfigPolicy;
+import org.optaplanner.core.impl.partitionedsearch.partitioner.SolutionPartitioner;
+import org.optaplanner.core.impl.phase.AbstractPhaseFactory;
+import org.optaplanner.core.impl.solver.recaller.BestSolutionRecaller;
+import org.optaplanner.core.impl.solver.termination.Termination;
+import org.optaplanner.core.impl.solver.thread.ChildThreadType;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class DefaultPartitionedSearchPhaseFactory<Solution_>
+        extends AbstractPhaseFactory<Solution_, PartitionedSearchPhase<Solution_>, PartitionedSearchPhaseConfig> {
+
+    private static final Logger logger = LoggerFactory.getLogger(DefaultPartitionedSearchPhaseFactory.class);
+
+    public DefaultPartitionedSearchPhaseFactory(
+            PartitionedSearchPhaseConfig phaseConfig) {
+        super(phaseConfig);
+    }
+
+    @Override
+    public PartitionedSearchPhase<Solution_> buildPhase(int phaseIndex, HeuristicConfigPolicy solverConfigPolicy,
+            BestSolutionRecaller<Solution_> bestSolutionRecaller, Termination solverTermination) {
+        HeuristicConfigPolicy phaseConfigPolicy = solverConfigPolicy.createPhaseConfigPolicy();
+        ThreadFactory threadFactory = solverConfigPolicy.buildThreadFactory(ChildThreadType.PART_THREAD);
+        DefaultPartitionedSearchPhase phase = new DefaultPartitionedSearchPhase(
+                phaseIndex, solverConfigPolicy.getLogIndentation(), bestSolutionRecaller,
+                buildPhaseTermination(phaseConfigPolicy, solverTermination),
+                buildSolutionPartitioner(), threadFactory, resolvedActiveThreadCount(phaseConfig.getRunnablePartThreadLimit()));
+        List<PhaseConfig> phaseConfigList_ = phaseConfig.getPhaseConfigList();
+        if (ConfigUtils.isEmptyCollection(phaseConfigList_)) {
+            phaseConfigList_ = Arrays.asList(
+                    new ConstructionHeuristicPhaseConfig(),
+                    new LocalSearchPhaseConfig());
+        }
+        phase.setPhaseConfigList(phaseConfigList_);
+        phase.setConfigPolicy(phaseConfigPolicy.createChildThreadConfigPolicy(ChildThreadType.PART_THREAD));
+        EnvironmentMode environmentMode = phaseConfigPolicy.getEnvironmentMode();
+        if (environmentMode.isNonIntrusiveFullAsserted()) {
+            phase.setAssertStepScoreFromScratch(true);
+        }
+        if (environmentMode.isIntrusiveFastAsserted()) {
+            phase.setAssertExpectedStepScore(true);
+            phase.setAssertShadowVariablesAreNotStaleAfterStep(true);
+        }
+        return phase;
+    }
+
+    private SolutionPartitioner buildSolutionPartitioner() {
+        if (phaseConfig.getSolutionPartitionerClass() != null) {
+            SolutionPartitioner<?> solutionPartitioner = ConfigUtils.newInstance(this,
+                    "solutionPartitionerClass", phaseConfig.getSolutionPartitionerClass());
+            ConfigUtils.applyCustomProperties(solutionPartitioner, "solutionPartitionerClass",
+                    phaseConfig.getSolutionPartitionerCustomProperties(), "solutionPartitionerCustomProperties");
+            return solutionPartitioner;
+        } else {
+            if (phaseConfig.getSolutionPartitionerCustomProperties() != null) {
+                throw new IllegalStateException(
+                        "If there is no solutionPartitionerClass (" + phaseConfig.getSolutionPartitionerClass()
+                                + "), then there can be no solutionPartitionerCustomProperties ("
+                                + phaseConfig.getSolutionPartitionerCustomProperties() + ") either.");
+            }
+            // TODO Implement generic partitioner
+            throw new UnsupportedOperationException();
+        }
+    }
+
+    protected Integer resolvedActiveThreadCount(String runnablePartThreadLimit) {
+        int availableProcessorCount = getAvailableProcessors();
+        Integer resolvedActiveThreadCount;
+        final boolean threadLimitNullOrAuto =
+                runnablePartThreadLimit == null || runnablePartThreadLimit.equals(ACTIVE_THREAD_COUNT_AUTO);
+        if (threadLimitNullOrAuto) {
+            // Leave one for the Operating System and 1 for the solver thread, take the rest
+            resolvedActiveThreadCount = Math.max(1, availableProcessorCount - 2);
+        } else if (runnablePartThreadLimit.equals(ACTIVE_THREAD_COUNT_UNLIMITED)) {
+            resolvedActiveThreadCount = null;
+        } else {
+            resolvedActiveThreadCount = ConfigUtils.resolveThreadPoolSizeScript("runnablePartThreadLimit",
+                    runnablePartThreadLimit, ACTIVE_THREAD_COUNT_AUTO, ACTIVE_THREAD_COUNT_UNLIMITED);
+            if (resolvedActiveThreadCount < 1) {
+                throw new IllegalArgumentException("The runnablePartThreadLimit (" + runnablePartThreadLimit
+                        + ") resulted in a resolvedActiveThreadCount (" + resolvedActiveThreadCount
+                        + ") that is lower than 1.");
+            }
+            if (resolvedActiveThreadCount > availableProcessorCount) {
+                logger.debug("The resolvedActiveThreadCount ({}) is higher than "
+                        + "the availableProcessorCount ({}), so the JVM will "
+                        + "round-robin the CPU instead.", resolvedActiveThreadCount, availableProcessorCount);
+            }
+        }
+        return resolvedActiveThreadCount;
+    }
+
+    protected int getAvailableProcessors() {
+        return Runtime.getRuntime().availableProcessors();
+    }
+}
