+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/phase/NoChangePhaseFactory.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.optaplanner.core.impl.phase;
+
+import org.optaplanner.core.config.phase.NoChangePhaseConfig;
+import org.optaplanner.core.impl.heuristic.HeuristicConfigPolicy;
+import org.optaplanner.core.impl.solver.recaller.BestSolutionRecaller;
+import org.optaplanner.core.impl.solver.termination.Termination;
+
+public class NoChangePhaseFactory<Solution_>
+        extends AbstractPhaseFactory<Solution_, NoChangePhase<Solution_>, NoChangePhaseConfig> {
+
+    public NoChangePhaseFactory(NoChangePhaseConfig phaseConfig) {
+        super(phaseConfig);
+    }
+
+    @Override
+    public NoChangePhase<Solution_> buildPhase(int phaseIndex, HeuristicConfigPolicy solverConfigPolicy,
+            BestSolutionRecaller bestSolutionRecaller, Termination solverTermination) {
+        HeuristicConfigPolicy phaseConfigPolicy = solverConfigPolicy.createPhaseConfigPolicy();
+        return new NoChangePhase(
+                phaseIndex, solverConfigPolicy.getLogIndentation(), bestSolutionRecaller,
+                buildPhaseTermination(phaseConfigPolicy, solverTermination));
+    }
+}
