+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/phase/PhaseFactoryProvider.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.optaplanner.core.impl.phase;
+
+import org.optaplanner.core.config.constructionheuristic.ConstructionHeuristicPhaseConfig;
+import org.optaplanner.core.config.exhaustivesearch.ExhaustiveSearchPhaseConfig;
+import org.optaplanner.core.config.localsearch.LocalSearchPhaseConfig;
+import org.optaplanner.core.config.partitionedsearch.PartitionedSearchPhaseConfig;
+import org.optaplanner.core.config.phase.NoChangePhaseConfig;
+import org.optaplanner.core.config.phase.PhaseConfig;
+import org.optaplanner.core.config.phase.custom.CustomPhaseConfig;
+import org.optaplanner.core.impl.constructionheuristic.DefaultConstructionHeuristicPhaseFactory;
+import org.optaplanner.core.impl.exhaustivesearch.DefaultExhaustiveSearchPhaseFactory;
+import org.optaplanner.core.impl.localsearch.DefaultLocalSearchPhaseFactory;
+import org.optaplanner.core.impl.partitionedsearch.DefaultPartitionedSearchPhaseFactory;
+import org.optaplanner.core.impl.phase.custom.DefaultCustomPhaseFactory;
+
+public class PhaseFactoryProvider<Solution_> {
+
+    public AbstractPhaseFactory createPhaseFactory(PhaseConfig<?> phaseConfig) {
+        if (LocalSearchPhaseConfig.class.isAssignableFrom(phaseConfig.getClass())) {
+            return new DefaultLocalSearchPhaseFactory<Solution_>((LocalSearchPhaseConfig) phaseConfig);
+        } else if (ConstructionHeuristicPhaseConfig.class.isAssignableFrom(phaseConfig.getClass())) {
+            return new DefaultConstructionHeuristicPhaseFactory<Solution_>((ConstructionHeuristicPhaseConfig) phaseConfig);
+        } else if (PartitionedSearchPhaseConfig.class.isAssignableFrom(phaseConfig.getClass())) {
+            return new DefaultPartitionedSearchPhaseFactory<Solution_>((PartitionedSearchPhaseConfig) phaseConfig);
+        } else if (CustomPhaseConfig.class.isAssignableFrom(phaseConfig.getClass())) {
+            return new DefaultCustomPhaseFactory<Solution_>((CustomPhaseConfig) phaseConfig);
+        } else if (ExhaustiveSearchPhaseConfig.class.isAssignableFrom(phaseConfig.getClass())) {
+            return new DefaultExhaustiveSearchPhaseFactory<Solution_>((ExhaustiveSearchPhaseConfig) phaseConfig);
+        } else if (NoChangePhaseConfig.class.isAssignableFrom(phaseConfig.getClass())) {
+            return new NoChangePhaseFactory<Solution_>((NoChangePhaseConfig) phaseConfig);
+        } else {
+            throw new IllegalArgumentException(
+                    String.format("Unknown PhaseConfig type: (%s).", phaseConfig.getClass().getName()));
+        }
+    }
+}
